Rename rhombus text package
