+++ app/src/main/java/com/skul/yuriy/composeplayground/feature/sensorRotation/text/rhombustext/RhombusText.kt
@@ -1,4 +1,4 @@
-package com.skul.yuriy.composeplayground.feature.sensorRotation.text.rhombus
+package com.skul.yuriy.composeplayground.feature.sensorRotation.text.rhombustext
 
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
